remove old locations component
