+++ src/components/Loc/index.js
@@ -1,21 +1,43 @@
 import React, { useState, useContext, useEffect } from "react";
 import { useStaticQuery, graphql } from "gatsby";
-import { Title, H1, H2, H3, H4, Span, Paragraph, Separator } from "../Heading";
+import { Title, H1, H2, H3, H4, Span, Paragraph } from "../Heading";
 import { GridContainer, Grid, Div } from "../Sections";
-import {
-  Button,
-  Colors,
-  RoundImage,
-  StyledBackgroundSection,
-} from "../Styling";
-import Icon from "../Icon";
-import dayjs from "dayjs";
-import "dayjs/locale/de";
+import { Img, Colors } from "../Styling";
+
 import { getCohorts } from "../../actions";
 import { SessionContext } from "../../session.js";
 import Link from "gatsby-link";
 
-const Loc = ({ locations, title, paragraph, lang }) => {
+const Loc = ({ lang, yml, allLocationYaml }) => {
+  const { heading, image, sub_heading, choose, regions } = yml;
+
+  useEffect(() => {
+    regions.forEach((reg, ind, arr) => {
+      if (arr[ind].name === "online") {
+        arr[ind].sub_links = allLocationYaml.edges.filter(
+          (loc) =>
+            loc.node.online_available || loc.node.online_available === null
+        );
+      } else {
+        arr[ind].sub_links = allLocationYaml.edges.filter(
+          (loc) => loc.node.meta_info.region === reg.name || (
+            reg.name === 'america' && (loc.node.meta_info.region === 'usa-canada' || loc.node.meta_info.region === 'latam'))
+        );
+      }
+
+      arr[ind].sub_links.sort((a, b) => {
+        if (a.node.meta_info.position < b.node.meta_info.position) {
+          return -1;
+        }
+        if (a.node.meta_info.position > b.node.meta_info.position) {
+          return 1;
+        }
+        return 0;
+      });
+    });
+    setActiveOpt({ ...regions[0] });
+  }, []);
+
   const data = useStaticQuery(graphql`
     {
       allLocYaml {
@@ -35,45 +57,14 @@ const Loc = ({ locations, title, paragraph, lang }) => {
   );
   if (content) content = content.node;
   else return null;
-  const { session } = useContext(SessionContext);
-  const [index, setIndex] = useState(null);
-  const [status, setStatus] = useState({ toggle: false, hovered: false });
-  const [datas, setData] = useState({
-    cohorts: { catalog: [], all: [], filtered: [] },
+
+  const [activeOpt, setActiveOpt] = useState({
+    ...regions[0],
   });
-  useEffect(() => {
-    const getData = async () => {
-      const cohorts = await getCohorts();
-      let _types = [];
-      setData((oldData) => ({
-        cohorts: {
-          catalog: oldData.cohorts.catalog,
-          all: cohorts,
-          filtered: cohorts,
-        },
-      }));
-    };
-    getData();
-  }, []);
-  let loc = locations
-    .filter(
-      (l) => !["unlisted", "hidden"].includes(l.node.meta_info.visibility)
-    )
-    .sort((a, b) =>
-      a.node.meta_info.position > b.node.meta_info.position ? 1 : -1
-    );
-  const nextDate = (location) => {
-    let cohort = datas.cohorts.all.find(
-      (item) => item.academy.slug === location.node.breathecode_location_slug
-    );
-    let onlineCohort = datas.cohorts.all.find(
-      (item) => item.academy.slug === "online"
-    );
-    return cohort || onlineCohort;
-  };
+
   return (
     <>
-      {title && (
+      {heading && (
         <GridContainer
           margin_tablet="0 0 35px 0"
           margin="0 0 32px 0"
@@ -86,87 +77,191 @@ const Loc = ({ locations, title, paragraph, lang }) => {
               lineHeight="19px"
               fontWeight="900"
             >
-              {title}
+              {heading}
             </H2>
-            <Paragraph>{paragraph}</Paragraph>
+            <Paragraph>{sub_heading}</Paragraph>
           </Div>
         </GridContainer>
       )}
-      <GridContainer
-        columns="1"
-        columns_sm="2"
-        columns_tablet="3"
-        gridGap="0"
-        columnGap="15px"
-        margin="0 0 50px 0"
-        margin_tablet="0 0 70px 0"
+      <hr
+        style={{
+          margin: "auto",
+          width: "80%",
+          marginBottom: "20px",
+          border: `1px solid ${Colors.verylightGray}`,
+        }}
+      />
+      <Div
+        id="locations-container"
+        padding="0 10% 10% 10%"
+        flexDirection_tablet="row"
+        flexDirection_sm="column"
+        flexDirection_xs="column"
       >
-        {loc !== null &&
-          loc.map((item, i) => {
-            const next = nextDate(item);
-            let stringDate = "";
-            if (next !== undefined && next.kickoff_date) {
-              stringDate = dayjs(next.kickoff_date)
-                .locale(lang)
-                .format("ddd DD MMM YYYY");
-              stringDate =
-                stringDate[0].toUpperCase() +
-                stringDate.substr(1, 2) +
-                "." +
-                stringDate.substr(3);
-            }
-            return (
-              <Div
-                onMouseOver={() => setIndex(i)}
-                key={i}
-                style={{
-                  borderBottom: `1px solid ${Colors.lightGray}`,
-                  position: "relative",
-                  transition: "background 0.3s ease, border-left 0.3s ease",
-                  borderLeft: `${
-                    index === i ? "10px solid" + Colors.yellow : "0"
-                  }`,
-                }}
-                display="flex"
-                flexDirection="row"
-                justifyContent="between"
-                // height="207px"
-                height="auto"
-                padding="30px 24px"
-                background={index === i ? Colors.verylightGray : Colors.white}
-              >
-                <H3 textAlign="left">
-                  {item.node.name}
-                  <Span animated color={Colors.yellow}>
-                    _
-                  </Span>
-                </H3>
-                {/* <Div
+        {image && (
+          <Div
+            id="img-container"
+            width_tablet="30%"
+            maxWidth_tablet="270px"
+            width_xs="100%"
+            margin_tablet="0 20px 0 0"
+            margin_xs="0 0 20px 0"
+            // height="100%"
+          >
+            <Img
+              src={image}
+              // borderRadius={"1.25rem"}
+              borderRadius={"3px"}
+              // className="pointer"
+              alt={"4Geeks Academy Section"}
+              margin="auto"
+              width="100%"
+              height="100%"
+              minHeight_tablet="none"
+              minHeight_sm="100px"
+              backgroundSize={`cover`}
+              style={{ imageRendering: "crisp-edges" }}
+            />
+          </Div>
+        )}
+        <Div
+          id="menu-container"
+          width_tablet="70%"
+          width_xs="100%"
+          display="block"
+          // flexShrink_tablet="0"
+        >
+          {choose && (
+            <Paragraph
+              textAlign="left"
+              color={Colors.darkGray}
+              margin="0 0 10px 0"
+            >
+              {choose}
+            </Paragraph>
+          )}
+          <Div
+            id="selectors-container"
+            flexDirection_tablet="row"
+            flexDirection_xs="column"
+            width_xs="100%"
+            // flexShrink_tablet="0"
+          >
+            <Div
+              id="options-container"
+              flexDirection_tablet="column"
+              justifyContent_tablet="start"
+              flexDirection_xs="row"
+              justifyContent_xs="between"
+              width_tablet="33%"
+              width_xs="100%"
+            >
+              {regions.map((m, i) => (
+                <Div
+                  color={
+                    activeOpt.title === m.title ? Colors.black : Colors.gray
+                  }
+                  borderLeft_tablet={
+                    activeOpt.title === m.title
+                      ? `5px solid ${Colors.blue}`
+                      : null
+                  }
+                  borderBottom_tablet={"none"}
+                  borderLeft_xs={"none"}
+                  borderBottom_xs={
+                    activeOpt.title === m.title
+                      ? `5px solid ${Colors.blue}`
+                      : null
+                  }
+                  borderRadius="none"
+                  padding="10px"
+                  onClick={() => {
+                    setActiveOpt({ ...m });
+                  }}
+                  style={{ cursor: "pointer" }}
                   display="block"
-                  display_tablet="block"
                 >
-                  <Paragraph textAlign="left" fontSize="15px" lineHeight="22px" color={Colors.darkGray}>
-                    {content.label}
-                  </Paragraph >
-                  <Paragraph textAlign="left" fontSize="15px" lineHeight="22px" color={Colors.darkGray}>
-                    {next !== undefined ? next.syllabus_version?.name : "No upcoming dates at this location"}
+                  <H3
+                    textAlign="left"
+                    fontSize="20px"
+                    color={
+                      activeOpt.title === m.title ? Colors.black : Colors.gray
+                    }
+                  >
+                    {m.title}
+                  </H3>
+                  <Paragraph
+                    textAlign="left"
+                    display_tablet="block"
+                    display_xs="none"
+                    display_sm="none"
+                    color={
+                      activeOpt.title === m.title ? Colors.black : Colors.gray
+                    }
+                  >
+                    {m.content}
                   </Paragraph>
-                  {next !== undefined && next.kickoff_date && <Paragraph textAlign="left" fontSize="15px" lineHeight="22px" color={Colors.darkGray}>
-                    <span className="capitalize">{stringDate}</span>
-                  </Paragraph>}
-                </Div> */}
-                <Link to={`/${lang}/coding-campus/${item.node.meta_info.slug}`}>
-                  <Icon
-                    // style={{position: "absolute", bottom: "18px", right: "18px"}}
-                    icon="arrowright"
-                    height="32px"
-                    width="32px"
-                  />
-                </Link>
-              </Div>
-            );
-          })}
-      </GridContainer>
+                </Div>
+              ))}
+            </Div>
+            <Paragraph
+              display_tablet="none"
+              display_xs="block"
+              textAlign="left"
+              color={Colors.darkGray}
+              margin="10px 0"
+            >
+              {activeOpt.content}
+            </Paragraph>
+            <Div
+              id="links-container"
+              // flexShrink_tablet="0"
+              flexDirection="column"
+              width_tablet="67%"
+              width_xs="100%"
+              flexWrap="wrap"
+              // maxHeight="330px"
+              // minHeight="330px"
+              height="330px"
+              alignContent_tablet="flex-start"
+              alignContent_xs="space-between"
+            >
+              {activeOpt.sub_links != undefined &&
+                Array.isArray(activeOpt.sub_links) &&
+                activeOpt.sub_links.map((l, i) => {
+                  return (
+                    <Div
+                      margin_tablet="2px 10px 2px 60px"
+                      margin_xs="2px 0"
+                      padding="10px 0 0 0"
+                      alignItems="baseline"
+                      maxWidth_tablet="none"
+                      maxWidth_xs="50%"
+                    >
+                      <Link
+                        to={`/${lang}/coding-campus/${l.node.meta_info.slug}`}
+                        key={i}
+                        // style={{maxWidth:'50%'}}
+                      >
+                        <H3
+                          textAlign="left"
+                          width="fit-content"
+                          fontSize="15px"
+                          lineHeight="20px"
+                          fontWeight="400"
+                          margin="0 5px 0 0"
+                          borderBottomHover="2px solid black"
+                        >
+                          {l.node.name}
+                        </H3>
+                      </Link>
+                    </Div>
+                  );
+                })}
+            </Div>
+          </Div>
+        </Div>
+      </Div>
     </>
   );
 };
